Postpone termination of service during system reboot
This change is necessary for the correct EVS application exit.

Below a detailed explanation.
When the system reboots the init process asks all processes to stop
in the following way:
 - first it notifies the normal processes via SIGTERM (and if they
   don't stop via SIGKILL);
 - next it notifies in the same way the processes marked as a `shutdown
   critical`.
The order in which the init stops processes inside one group (normal or
critical) is undefined.

As result, such behavior may lead case when one of two interacting
processes is stopped and another process still tries to interact with
the first one.

There is no way how to explain to the init process about dependencies
between processes. Only one way how to fix the issue is marking one of
the processes as `shutdown critical` that allows postponing its
termination.

See for details:
 - $ANDROID_BUILD_TOP/system/core/init/README.md
 - $ANDROID_BUILD_TOP/system/core/init/reboot.cpp

For debug add `printk.devkmsg=on` to bootargs, reboot DUT and check a
minicom output.

Test: reboot and ensure that logcat doesn't contain messages like:
      ...
      signal 6 (SIGABRT), code -1 (SI_QUEUE), fault addr
      ...
      Abort message: 'Failed HIDL return status not checked. Usually
      this happens because of a transport error (error parceling, binder
      driver, or from unparceling). If you see this in code calling into
      "Bn" classes in for a HAL server process, then it is likely that
      the code there is returning transport errors there (as opposed to
      errors defined within its protocol). Error is:
      Status(EX_TRANSACTION_FAILED): 'DEAD_OBJECT:
      ...
      android.hardware.automotive.evs@1.0.so
      ...
      android::hardware::automotive::evs::V1_0::BnHwEvsDisplay
      ...
      _hidl_returnTargetBufferForDisplay(android::hidl::base::V1_0
Jira-Id: ADM-5092
Change-Id: Idd18995ebfa1d2d9e0c19267a3d5f863299f0721
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>